Strong/weak solution for Cat/Spot done. Un/archiving works correctly. Final solution is to unarchive savedSpots (for spots with nil categories); then do categories and for each category's spot, create the spot, link it to the category and replace this spot in savedSpots (use savedSpotIndex)

+++ FoodAlert/FoodAlert/Spot.h
@@ -24,7 +24,7 @@
 @property (nonatomic) BOOL saved;
 
 @property (weak, nonatomic) Categorie* category; // Category owns array of Spots, so weak link back
-@property (nonatomic) short savedSpotIndex;
+@property (nonatomic) int savedSpotIndex;
 
 @property (nonatomic) NSString* notes;
